pyproject.toml
[tool.poetry]
name = "gcp"
version = "0.1.0"
description = "CLI tool automating administrative tasks in GCP like creating POSIX users and managing their SSH keys"
authors = ["Viacheslav Bessonov <viacheslav.bessonov@algalon.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-api-python-client = "^1.7.11"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
click = "^7.0"
cryptography = "^2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
